Switch CI to Ubuntu bionic to use OpenSSL 1.1.x.
The newest version of OpenSSL available on Ubuntu Xenial is 1.0.2,
which doesn't support the expired AddTrust certificate even when
another certificate in the chain is trusted as a root. This makes
the cert verification fail for the OIDC conformance tests, since the
server presents a chain that includes the expired cert but *should*
validate successfully (and does in newer OpenSSL versions).