Skip unrecognized algorithms and keys in provider metadata
Previously, this crate would return a deserialization error if it encountered
any unrecognized signature or encryption algorithms in the `ProviderMetadata`
and `JsonWebKeySet`. This patch changes the behavior to instead ignore unexpected
values, with the aim of improving compatibility with OIDC providers that support
other algorithms but don't necessarily use them to sign JWTs.

This change focuses on the signing/encryption enums since those are difficult
to override via type parameters due to the interdependence between the key
and algorithm traits/types.