Take JWK alg into account during key selection (#131)
RFC 7517 (Sec. 4.5) permits multiple keys in a JWK set to have the same
key ID. In such a JWK set, the `alg` (algorithm) JWK field is useful
for identifying the key that should be used to verify the signature of
a JWT.

Previously, this crate did not examine the `alg` JWK field, leading to
`AmbiguousKeyId` errors in cases where multiple keys in a JWK set
shared a key ID. This change filters out keys with `alg` fields that
differ from the `alg` field in the JWT JOSE header, resolving the
ambiguity in many cases.

To avoid breaking changes (primarily to the `JsonWebKey` trait), this
behavior is behind a new `jwk-alg` feature flag. This feature flag
will be removed in a future SemVer major version.