add support for specific JOSE header types
Check https://tools.ietf.org/html/rfc7519#section-5.1 and https://tools.ietf.org/html/rfc7515#section-4.1.9 for more information on when the typ header should be set and to which value.

This commit allows to skip the check altogether as warranted by RFC 7515 the use of the header is optional but if set, according to RFC 7519, it should be set to "JWT". For other token types this is different (see RFC 9068).