Adds `IdTokenVerifier::require_typ_check` method
This crate validates the `typ` field of the JOSE header equals `"JWT"`
case-insensitively.

However, some identity providers return tokens with `typ` values such
`"at+jwt"` which are largely compatible but impossible to use with this
crate.

This change introduces a builder method on `IdTokenVerifier` which
allows disabling this check for users that wish to.

The current behavior is retained as the default (i.e. it requires the
value to equal `"JWT"`).